+++ pulse2percept/utils/__init__.py
@@ -18,7 +18,8 @@
 """
 from .base import (PrettyPrint, FreezeError, Frozen, Parametrized, Data,
                    bijective26_name, cached, gamma)
-from .animation import HTMLAnimation, frame_interval
+from .animation import (FrameTimeline, HTMLAnimation, frame_interval,
+                        frame_timeline)
 from .geometry import (cart2pol, pol2cart, delta_angle)
 from .array import is_strictly_increasing, radial_mask, sample, unique
 from .images import center_image, scale_image, shift_image, trim_image
@@ -45,6 +46,8 @@ __all__ = [
     'deprecated',
     'deprecated_alias',
     'frame_interval',
+    'frame_timeline',
+    'FrameTimeline',
     'FreezeError',
     'Frozen',
     'gamma',